+++ packages/stylelint-config-suitcss/__tests__/index.js
@@ -84,12 +84,12 @@ const invalidCss = (
 test("no warnings with valid css", t => {
   return stylelint.lint({
     code: validCss,
-    config: config,
+    config,
   })
   .then(data => {
     const { errored, results } = data
     const { warnings } = results[0]
-    t.notOk(errored, "no errored")
+    t.falsy(errored, "no errored")
     t.is(warnings.length, 0, "flags no warnings")
   })
 })
@@ -97,12 +97,12 @@ test("no warnings with valid css", t => {
 test("a warning with invalid css", t => {
   return stylelint.lint({
     code: invalidCss,
-    config: config,
+    config,
   })
   .then(data => {
     const { errored, results } = data
     const { warnings } = results[0]
-    t.ok(errored, "errored")
+    t.truthy(errored, "errored")
     t.is(warnings.length, 1, "flags one warning")
     t.is(warnings[0].text, "Expected a leading zero (number-leading-zero)", "correct warning text")
   })