+++ packages/stylelint-config-suitcss/CHANGELOG.md
@@ -1,3 +1,13 @@
+# Head
+
+* Updated: `ava` to `0.18.2`
+* Updated: `eslint` to `3.17.0`
+* Updated: `eslint-config-stylelint` to `6.0.0`
+* Updated: `stylelint-order` to `0.3.0`
+* Added: `suitcss/root-no-standard-properties`
+* Added: `suitcss/selector-root-no-composition`
+* Added: `suitcss/custom-property-no-outside-root`
+
 # 9.0.0
 
 * Removed: `root-no-standard-properties`.